separate grad
