+++ src/rgb_led/mod.rs
@@ -1,3 +1,8 @@
-mod led_sequence;
+mod gradient;
+mod math;
+mod plugin;
+mod sequence;
 
-pub use led_sequence::*;
+pub use gradient::*;
+pub use plugin::RgbLedPlugin;
+pub use sequence::*;
